requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
openai==1.99.9
python-dotenv==1.0.0
